refactor: switch ui to skeleton
CU-8697eyefy
