+++ package.json
@@ -23,6 +23,8 @@
 		"@graphql-codegen/typescript-operations": "^4.4.0",
 		"@graphql-typed-document-node/core": "^3.2.0",
 		"@popperjs/core": "^2.11.8",
+		"@skeletonlabs/skeleton": "3.0.0-next.10",
+		"@skeletonlabs/skeleton-svelte": "1.0.0-next.17",
 		"@solana/web3.js": "^1.98.0",
 		"@sveltejs/adapter-static": "^3.0.8",
 		"@sveltejs/kit": "^2.15.1",
@@ -40,10 +42,9 @@
 		"eslint": "^9.17.0",
 		"eslint-config-prettier": "^9.1.0",
 		"eslint-plugin-svelte": "^2.46.1",
-		"flowbite": "^2.5.2",
-		"flowbite-svelte": "^0.47.4",
 		"globals": "^15.14.0",
 		"graphql": "^16.10.0",
+		"lucide-svelte": "^0.471.0",
 		"prettier": "^3.4.2",
 		"prettier-plugin-svelte": "^3.3.2",
 		"prettier-plugin-tailwindcss": "^0.6.9",
