+++ src/app.html
@@ -7,7 +7,7 @@
 		<!doctype html>
 		%sveltekit.head%
 	</head>
-	<body data-sveltekit-preload-data="hover">
+	<body data-sveltekit-preload-data="hover" data-theme="wintry">
 		<div style="display: contents">%sveltekit.body%</div>
 	</body>
 </html>
