+++ src/lib/features/navbar/authn/LoginButton.svelte
@@ -1,9 +1,10 @@
 <script lang="ts">
-	import { GradientButton } from 'flowbite-svelte';
 	import { initWeb3Auth } from '$lib/services/auth/web3auth';
 	import { generateNonce, login } from './mutations.gql';
 	import { SolanaClient } from '$lib/services/solana/client';
 	import { setRefreshToken, setSessionToken } from '$lib/storage/local';
+	import { LogIn } from 'lucide-svelte';
+	import { Navigation } from '@skeletonlabs/skeleton-svelte';
 
 	const handleLogin = async () => {
 		const web3auth = await initWeb3Auth();
@@ -30,4 +31,4 @@
 	};
 </script>
 
-<GradientButton size="sm" color="lime" onclick={handleLogin}>Login</GradientButton>
+<Navigation.Tile id="btn-login" labelExpanded="Login" title="login" onclick={handleLogin}><LogIn /></Navigation.Tile>
