+++ src/lib/features/navbar/authn/LogoutButton.svelte
@@ -1,7 +1,8 @@
 <script>
 	import { initWeb3Auth } from '$lib/services/auth/web3auth';
 	import { removeRefreshToken, removeSessionToken } from '$lib/storage/local';
-	import { GradientButton } from 'flowbite-svelte';
+	import { Navigation } from '@skeletonlabs/skeleton-svelte';
+	import { LogOut } from 'lucide-svelte';
 
 	const handleLogout = async () => {
 		const web3auth = await initWeb3Auth();
@@ -13,4 +14,6 @@
 	};
 </script>
 
-<GradientButton size="sm" outline color="lime" onclick={handleLogout}>Logout</GradientButton>
+<Navigation.Tile id="btn-logout" labelExpanded="Logout" title="logout" onclick={handleLogout}
+	><LogOut /></Navigation.Tile
+>
