+++ src/routes/+layout.svelte
@@ -16,6 +16,11 @@
 	setUserInfoContext(user);
 </script>
 
-<Navbar></Navbar>
-
-{@render children?.()}
+<div class="grid grid-cols-[auto_1fr]">
+	<aside class="card surface-100 sticky top-0 col-span-1 grid h-screen">
+		<Navbar />
+	</aside>
+	<main class="col-span-1">
+		{@render children?.()}
+	</main>
+</div>
