+++ src/routes/+page.svelte
@@ -1,1 +1,3 @@
-<div class="w-full pt-20">Welcome to Growteer!</div>
+<div class="p-4">
+	<h1 class="h1">Welcome to Growteer!</h1>
+</div>
